[moe refactor] fix moe zero test, and little bug in low level zero
