+++ colossalai/shardformer/modeling/mixtral.py
@@ -37,7 +37,7 @@ class EPMixtralSparseMoeBlock(MixtralSparseMoeBlock):
         self.expert_start_idx = self.ep_rank * self.num_experts_per_ep
         held_experts = self.experts[self.expert_start_idx : self.expert_start_idx + self.num_experts_per_ep]
         set_tensors_to_none(self.experts, exclude=set(held_experts))
-        for p in self.experts.parameters():
+        for n, p in self.experts.named_parameters():
             p.ep_group = ep_group
 
     @staticmethod
