+++ colossalai/zero/low_level/low_level_optim.py
@@ -133,7 +133,7 @@ class LowLevelZeroOptimizer(OptimizerWrapper):
             group_params = list()
             for param in param_group["params"]:
                 if param.requires_grad:
-                    if self._bucket_store.moe_extra_dp_pg is None:
+                    if self._bucket_store.moe_extra_dp_pg is not None:
                         # skip moe param
                         if is_moe_tensor(param):
                             self.working_moe_params.append(param)
@@ -161,7 +161,10 @@ class LowLevelZeroOptimizer(OptimizerWrapper):
                     param_group[key] = value
             self.master_moe_params = []
             for param in self.working_moe_params:
-                self.master_moe_params.append(param.clone().to(torch.float32).detach())
+                if self._master_weights:
+                    self.master_moe_params.append(param.clone().to(torch.float32).detach())
+                else:
+                    self.master_moe_params.append(param.detach())
             # create mapping from master to working for optimizer io
             self.moe_master_to_working_map = {}
             for master_moe_param, working_moe_param in zip(self.master_moe_params, self.working_moe_params):
@@ -622,7 +625,9 @@ class LowLevelZeroOptimizer(OptimizerWrapper):
                 grads = self._grad_store.get_partitioned_gradients_by_param_id(group_id, id(working_param))
                 if len(grads) > 0:
                     # moe hybrid zero
-                    if self._bucket_store.moe_extra_dp_pg is not None and is_moe_tensor(working_param):
+                    if self._bucket_store.moe_extra_dp_pg is not None and is_moe_tensor(
+                        working_param
+                    ):  # TODO(@haze188): this code may be useless for next refactor
                         real_working_params[group_id].append(working_param)
                         if self._grad_store._partition_grads:
                             grad = grads
@@ -656,6 +661,7 @@ class LowLevelZeroOptimizer(OptimizerWrapper):
 
         # update param for moe ep
         # move grad to master param and compute norm
+
         if len(self.working_moe_params) > 0:
             moe_grads = []
             for master_moe_param, working_moe_param in zip(self.master_moe_params, self.working_moe_params):
@@ -685,6 +691,7 @@ class LowLevelZeroOptimizer(OptimizerWrapper):
         if len(self.working_moe_params) > 0:
             for master_moe_param, working_moe_param in zip(self.master_moe_params, self.working_moe_params):
                 master_moe_param.grad = None
+
                 working_moe_param.data = (
                     master_moe_param.data.to(working_moe_param.device).to(working_moe_param.dtype).detach()
                 )
