+++ tests/test_moe/test_moe_zero_fwd_bwd.py
@@ -1,78 +1,121 @@
+from copy import deepcopy
+
 import pytest
 import torch
+import torch.distributed as dist
 
 import colossalai
-from colossalai.booster import Booster
-from colossalai.booster.plugin import LowLevelZeroPlugin
-from colossalai.moe.manager import MOE_MANAGER
-from colossalai.testing import rerun_if_address_is_in_use, spawn
+from colossalai.booster.plugin.moe_hybrid_parallel_plugin import MoeHybridParallelPlugin
+from colossalai.testing import parameterize, rerun_if_address_is_in_use, spawn
 from colossalai.testing.random import seed_all
-from tests.test_moe.moe_utils import MoeModel, delete_moe_info, run_fwd_bwd, sync_local_from_ep
+from colossalai.zero import LowLevelZeroOptimizer
+from tests.test_moe.moe_utils import MoeModel, loose_close
 
 
-def run_zero_test(local_rank, stage=1):
-    criterion = torch.nn.CrossEntropyLoss()
-
-    MOE_MANAGER.__init__()
-    MOE_MANAGER.setup(parallel="EP")
-    moe_model = MoeModel().bfloat16()
-    moe_optimizer = torch.optim.Adam(moe_model.parameters())
-    moe_plugin = LowLevelZeroPlugin(stage=stage, precision="bf16")
-    moe_booster = Booster(plugin=moe_plugin)
-    moe_model, moe_optimizer, _, _, _ = moe_booster.boost(moe_model, moe_optimizer)
-
-    MOE_MANAGER.__init__()
-    MOE_MANAGER.setup(parallel=None)
-    zero_model = MoeModel().bfloat16()
-    delete_moe_info(zero_model)
-    zero_optimizer = torch.optim.Adam(zero_model.parameters())
-    zero_plugin = LowLevelZeroPlugin(stage=stage, precision="bf16")
-    zero_booster = Booster(plugin=zero_plugin)
-    zero_model, zero_optimizer, _, _, _ = zero_booster.boost(zero_model, zero_optimizer)
-    sync_local_from_ep(zero_model, moe_model)
-
-    data = torch.randn(16, 4).bfloat16().cuda()
-    label = torch.randint(0, 4, (16,)).cuda()
-
-    zero_out = run_fwd_bwd(zero_model, data, label, criterion, zero_optimizer)
-    moe_out = run_fwd_bwd(moe_model, data, label, criterion, moe_optimizer)
-    assert torch.allclose(zero_out, moe_out)
-
-    for (moe_name, moe_param), (zero_name, zero_param) in zip(
-        moe_model.module.named_parameters(), zero_model.module.named_parameters()
-    ):
-        assert moe_name == zero_name
-        moe_grad_list = moe_optimizer._grad_store.get_partitioned_gradients_by_param_id(0, id(moe_param))
-        zero_grad_list = zero_optimizer._grad_store.get_partitioned_gradients_by_param_id(0, id(zero_param))
-        if hasattr(moe_param, "moe_info"):
-            assert len(moe_grad_list) == 0
-            if stage == 1:
-                zero_grad = zero_grad_list[local_rank].view(moe_param.grad.shape)
-            else:
-                zero_grad = zero_grad_list[0].view(moe_param.grad.shape)
-            assert torch.allclose(
-                moe_param.grad, zero_grad, atol=1e-5
-            ), f"zero grad:\n{moe_param.grad}\ntorch grad:\n{zero_grad}\nmax diff: {(moe_param.grad - zero_grad).abs().max()}, mean diff: {(moe_param.grad - zero_grad).abs().mean()}"
-        else:
-            assert len(moe_grad_list) > 0
-            assert len(moe_grad_list) == len(zero_grad_list)
-            for moe_grad, zero_grad in zip(moe_grad_list, zero_grad_list):
-                assert torch.allclose(moe_grad, zero_grad)
+def split_ddp_grad(grad, world_size):
+    with torch.no_grad():
+        grad = grad.clone().detach().flatten()
+        padding_size = (world_size - grad.numel() % world_size) % world_size
+        if padding_size > 0:
+            grad = torch.nn.functional.pad(grad, [0, padding_size])
+        splited_grad = grad.split(grad.numel() // world_size)
+    return splited_grad
 
 
-def run_dist(rank, world_size, port, stage):
+tokens, n_experts = 7, 4
+hidden_size = 8
+top_k = 2
+
+
+# @parameterize("dtype", [torch.float16, torch.bfloat16])
+@parameterize("dtype", [torch.bfloat16])
+@parameterize("master_weights", [False])
+def run_zero_1_with_original_model(world_size, master_weights: bool, dtype: torch.dtype):
+    torch.distributed.get_rank()
+
+    torch.cuda.set_device(dist.get_rank())
+
+    plugin = MoeHybridParallelPlugin(
+        precision="bf16",
+        tp_size=1,
+        pp_size=1,
+        ep_size=dist.get_world_size(),
+    )
+
+    seed_all(1453)
+    zero_model = MoeModel(ep_group=plugin.ep_group).cuda().to(dtype)
+
+    ori_model = deepcopy(zero_model).to(dtype)
+
+    zero_optimizer = torch.optim.SGD(zero_model.parameters(), lr=1)
+    zero_optimizer = LowLevelZeroOptimizer(
+        zero_optimizer,
+        overlap_communication=True,
+        initial_scale=1,
+        reduce_bucket_size=1024 * 1024,
+        master_weights=master_weights,
+        moe_extra_dp_process_group=plugin.ep_group,
+    )
+
+    ori_optimizer = torch.optim.SGD(ori_model.parameters(), lr=1)
+
+    # create
+    input_data = torch.rand(1, 4).cuda()
+
+    # zero-dp forward
+    zero_output = zero_model(input_data.to(dtype))
+
+    # torch-ddp forward
+    ori_output = ori_model(input_data.to(dtype))
+    loose_close(zero_output, ori_output, dtype=dtype)
+
+    # zero-dp backward
+    zero_optimizer.backward(zero_output.mean().float())
+
+    # torch-ddp backward
+    ori_output.mean().float().backward()
+
+    # check grad
+    for (n1, p1), (n2, p2) in zip(ori_model.named_parameters(), zero_model.named_parameters()):
+        if dist.get_rank() == 0:
+            print(n1, p1.shape, p1.grad is None, "\t", n2, p2.shape, p2.grad is None)
+
+        if p1.grad is not None:
+            if p2.grad is None:
+                zero_grad_list = zero_optimizer._grad_store.get_partitioned_gradients_by_param_id(1, id(p2))
+            else:  # moe param
+                loose_close(p1.grad, p2.grad, dtype=dtype)
+                continue
+
+            ori_grad_list = split_ddp_grad(
+                p1.grad, world_size
+            )  # just flatten the original model grad to match the zero model grad shape
+            for zero_grad, torch_grad in zip(zero_grad_list, ori_grad_list):
+                loose_close(zero_grad, torch_grad, dtype=dtype)
+
+    # zero-dp step
+    zero_optimizer.step()
+
+    # original model step
+    ori_optimizer.step()
+
+    # check updated param
+    for (n, p), z1p in zip(ori_model.named_parameters(), zero_model.parameters()):
+        loose_close(p.data, z1p.data, dtype=dtype)
+
+
+def run_dist(rank, world_size, port):
     colossalai.launch(rank=rank, world_size=world_size, host="localhost", port=port, backend="nccl")
-    seed_all(42 + rank)
-    run_zero_test(rank, stage=stage)
+    run_zero_1_with_original_model(world_size=world_size)
+    # run_zero_1_2()
 
 
 @pytest.mark.dist
 @pytest.mark.parametrize("world_size", [2])
-@pytest.mark.parametrize("stage", [1, 2])
 @rerun_if_address_is_in_use()
-def test_moe_zero_model(world_size, stage):
-    spawn(run_dist, world_size, stage=stage)
+def test_moe_zero_model(world_size):
+    spawn(run_dist, world_size)
 
 
 if __name__ == "__main__":
-    test_moe_zero_model(world_size=2, stage=1)
+    test_moe_zero_model(world_size=2)
